+++ packages/jsenv-plugin-toolbar/src/client/ui/settings/toolbar_animation_setting.js
@@ -2,7 +2,7 @@ import {
   animationsAreEnabled,
   enableAnimations,
   disableAnimations,
-} from "../core/toolbar_animation.js"
+} from "../../core/toolbar_animation.js"
 
 export const renderToolbarAnimationSetting = () => {
   const animCheckbox = document.querySelector("#toggle-anims")
