+++ packages/jsenv-plugin-toolbar/src/client/ui/settings/toolbar_theme_setting.js
@@ -1,0 +1,17 @@
+import {
+  getCurrentTheme,
+  switchToDefaultTheme,
+  switchToLightTheme,
+} from "../../core/toolbar_theme.js"
+
+export const renderToolbarThemeSetting = () => {
+  const checkbox = document.querySelector("#checkbox-dark-theme")
+  checkbox.checked = getCurrentTheme() === "dark"
+  checkbox.onchange = () => {
+    if (checkbox.checked) {
+      switchToDefaultTheme()
+    } else {
+      switchToLightTheme()
+    }
+  }
+}
